fix: cast offset_of(m).bytes to size_t in the array-interface descr
std::meta::offset_of(m).bytes is std::ptrdiff_t (signed); assigning it into a
std::size_t tripped -Werror=narrowing / -Werror=sign-conversion, which CI turns on
(WELDER_WERROR) but a local build does not — so the numpy-view commit built clean
locally yet broke the Linux/macOS/Windows build jobs. A field offset is non-negative,
so a static_cast is safe. Verified with a full -DWELDER_WERROR=ON build.

+++ src/welder/rods/python/array_interface.hpp
@@ -108,7 +108,10 @@ consteval std::size_t ai_entry_count(std::meta::info E) {
     std::size_t n{0}, running{0};
     for (std::meta::info m : std::meta::nonstatic_data_members_of(
              E, std::meta::access_context::unchecked())) {
-        const std::size_t off{std::meta::offset_of(m).bytes};
+        // offset_of(m).bytes is std::ptrdiff_t (signed); a field offset is
+        // non-negative, so cast to size_t (avoids -Werror=narrowing/sign-conversion).
+        const std::size_t off{
+            static_cast<std::size_t>(std::meta::offset_of(m).bytes)};
         if (off > running)
             ++n; // interior padding
         ++n;     // the field
@@ -131,7 +134,10 @@ ai_descr() {
     std::size_t i{0}, running{0};
     for (std::meta::info m : std::meta::nonstatic_data_members_of(
              E, std::meta::access_context::unchecked())) {
-        const std::size_t off{std::meta::offset_of(m).bytes};
+        // offset_of(m).bytes is std::ptrdiff_t (signed); a field offset is
+        // non-negative, so cast to size_t (avoids -Werror=narrowing/sign-conversion).
+        const std::size_t off{
+            static_cast<std::size_t>(std::meta::offset_of(m).bytes)};
         if (off > running)
             out[i++] = {"", std::define_static_string("|V" +
                                                       ai_uint_string(off - running))};
